Compare ints then..

+++ main.cpp
@@ -41,7 +41,7 @@ int main() {
   };
 
   "eigen"_test = [] {
-    Eigen::Matrix3f m;
+    Eigen::Matrix3i m;
     m << 1, 2, 3, 4, 5, 6, 7, 8, 9;
 
     expect(eq(m(3, 3), 9));
